[INJIMOB-3559]: [Release] Address code review comments (#11)
* [INJIMOB-3559]: Organize helper classes aligned with kotlin library


* [INJIMOB-3559]: Network Manager updates for Empty Response body


* [INJIMOB-3559] refactor: rename renderVC to generateCredentialDisplayContent


* [INJIMOB-3559] refactor: rename constants to camelcase


* [INJIMOB-3559] refactor: remove unused variable


* [INJIMOB-3559] refactor: extract function


* [INJIMOB-3559] refactor: remove unused function param


* [INJIMOB-3559] refactor: simplify handling multiple errors


* [INJIMOB-3559] refactor: remove unknown credential format


* [INJIMOB-3559] refactor: fix failing test


---------
