+++ Sources/InjiVcRenderer/constants/Constants.swift
@@ -1,39 +1,39 @@
 import Foundation
 
 struct Constants {
-    static let RENDER_METHOD = "renderMethod"
-    static let RENDER_SUITE = "renderSuite"
-    static let TEMPLATE = "template"
-    static let ID = "id"
-    static let TYPE = "type"
-    static let RENDER_PROPERTY = "renderProperty"
-    static let CREDENTIAL_SUBJECT = "credentialSubject"
+    static let renderMethod = "renderMethod"
+    static let renderSuite = "renderSuite"
+    static let template = "template"
+    static let id = "id"
+    static let type = "type"
+    static let renderProperty = "renderProperty"
+    static let credentialSubject = "credentialSubject"
 
-    static let DIGEST_MULTIBASE = "digestMultibase"
+    static let digestMultibase = "digestMultibase"
 
-    static let SVG_MUSTACHE = "svg-mustache"
-    static let TEMPLATE_RENDER_METHOD = "TemplateRenderMethod"
+    static let svgMustache = "svg-mustache"
+    static let templateRenderMethod = "TemplateRenderMethod"
 
-    static let DEFAULT_LOCALE = "eng"
+    static let defaultLocale = "eng"
 
-    static let ADDRESS_LINE_1 = "addressLine1"
-    static let ADDRESS_LINE_2 = "addressLine2"
-    static let ADDRESS_LINE_3 = "addressLine3"
-    static let CITY = "city"
-    static let PROVINCE = "province"
-    static let REGION = "region"
-    static let POSTAL_CODE = "postalCode"
+    static let addressLine1 = "addressLine1"
+    static let addressLine2 = "addressLine2"
+    static let addressLine3 = "addressLine3"
+    static let city = "city"
+    static let province = "province"
+    static let region = "region"
+    static let postalCode = "postalCode"
 
-    static let QR_CODE_PLACEHOLDER = "{{/qrCodeImage}}"
-    static let QR_CODE_IMAGE_ID = "qrCodeImage"
-    static let QR_CODE_FALLBACK_IMAGE_ID = "qrCodeFallbackImage"
+    static let qrCodePlaceholder = "{{/qrCodeImage}}"
+    static let qrCodeImageId = "qrCodeImage"
+    static let qrCodeFallbackImageId = "qrCodeFallbackImage"
 
-    static let QR_IMAGE_PREFIX = "data:image/png;base64"
-    static let CONCATENATED_ADDRESS = "concatenatedAddress"
-    static let CONCATENATED_ADDRESS_PLACEHOLDER = "{{/concatenatedAddress}}"
-    static let QRCODE_IMAGE_TYPE = "data:image/png;base64,"
+    static let qrImagePrefix = "data:image/png;base64"
+    static let concatenatedAddress = "concatenatedAddress"
+    static let concatenatedAddressPlaceholder = "{{/concatenatedAddress}}"
+    static let qrCodeImageType = "data:image/png;base64,"
 
-    static let UNKNOWN_ERROR = "Unknown Error"
+    static let unknownError = "Unknown Error"
 
-    static let FALLBACK_QR_CODE = "iVBORw0KGgoAAAANSUhEUgAAATIAAAE0CAYAAACvn7/YAAAACXBIWXMAACE4AAAhOAFFljFgAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAADaoSURBVHgB7Z1pkFxXledvLrUvKpVkLcZGZSNLNLaxkU1bot2S24NZJwJm6aEjiOiBGD5NdLuZIEwrpj9goukYh80H45lv4w7g0wARQ0NEN9M02G3JLPIAttxYxptMCdtabKtUUu1ZufT9n8zz6uat+7IyqyqXV/X/hZ7yZb79vbr/d865595rDCGEJJyUWSNHjx4dM4QQsk488MAD46ZBGhYyK1x3FYvFT9jZT5ZKpTFDCCHrz0lMqVTq+w8++OD3Vlq5LiG7//77R2ZnZ//CCtfn7dcRQwghrWPcTl9+6KGHvhG3wopCBgusUCh83c6OhZaPjFDXCCHrx+TkZNyicWtMffarX/3qE/6CmkL2xS9+8Ut2w/vd33p7e80dd9xh9uzZY3bt2iXfCSFkvZifnzfnz583L774onnhhRfM5cuXq5Zbd/N+625+ueq3uJ35IgbBOnLkiIgYIYS0ipMnT5pjx45VCZovZpnQhr6Iwfr69Kc/bfbu3WsIIaSVwPN797vfLVaaI2Z33XnnneanP/3pMXxZZpFVYmL/rN9vueUW84lPfMIQQki7+f73v2+effbZ6Ls1uP4IMbNlFtmhQ4cgYhLBhyX2qU99yhBCSCcAy2x8fDyyzKyLedeHPvShb1YJ2X333fcZ+4HJbNmyRUSMwXxCSCcxNjYmVlk+n8fXEfu5kPbW+ZLOILDP1ApCSKcBXTp8+HD03bqXfxEJ2Re+8IVPmkquGKyxW2+91RBCSCdy8OBB11sciYQsnU5HEX34oYQQ0sm4qWCuaxmZYPv37zeEENLJoDJSCQoZ8jYIIaSTcXVKhMztigd+J2sqCSGdDnRKKyTToYWEEJIk0oYQQhIOhYwQkngoZISQxEMhI4QkHgoZISTxUMgIIYmHQkYISTwUMkJI4qGQEUISD4WMEJJ4KGSEkMRDISOEJB4KGSEk8VDICCGJh0JGCEk8FDJCSOKhkBFCEg+FjBCSeChkhJDEQyEjhCQeChkhJPFQyAghiYdCRghJPBQyQkjioZARQhIPhYwQkngoZISQxJM1hKyCUqloTDFvPwv2s1j+LJXsVFhaHiCVqrw7Uxl8sf/sZzpd/rTLUmn+SZLG4V8NWRGIUqmwYEpFK1IFiFd+2ToiUDJBmDImFbuzYlnwgBXCYikX2FfW7sb+aVpRS1UmQmrBvxCyHCs0EK4iLC77GQlPxYJKZ/qWrCgITWptEQqx3qwlV7LHE+sOApefX1qhYqmlMt32sN1ybEJcKGSkjBWrYmHelPI5KySL5d9EuKyAdHVXLKMu0wzUmvP3j/OAqOk5lQo5A4cV66WyPRQ1EkEh2+RAIIq52SrxSmf7rFB0i4jhe7sQwYK42fPRcy3lraVoBa2Umy6LGqy0bK98ks0LhWwzAusrPyeTuI2ueDXJ6loPRNi6u6SqfUnUFkxhIScWXbqrn1baJoVCtpnwBEyEwbqN6UxvWy2v1RCJWmlAxKy4OGstS2ulQZQz1u3M9lPQNhEUss1AQMDSPf1Ntb6KC1PB39M9Q2ZdEWuyVyZ1k1FRgKn8OwVtM0Ah2+CIpbLOAoage2HhisnPTZjC7EWp2cQ8YldxAuYDQct0D0psK9u3zWT7t5m0/Z7t325WC64r07tFUjzK1z0v5ySCZt1OsnGhkG1QYJ0UICq2UK9VwCBci1Nn7XTOLE6fM3krXi6oQYQoZftGTXpwlxWkJatLLTBX4Iq5qXLAHvudPm9yk2eq9tc1tFvEDZ9dQ1c3HshHvAwiac8L90BFTWJoVtTIxoNCttGAG1kRCnG7bIFeTeGF8CxcPmNFZlwETIHl1LfjJpOxnxCbTA+sqh6zFsSim52wVt1FsfDwOffmczIBCFr3yJjpsROup14g3BBXscwqMTQk4dLd3HhQyDYQxbx17RanRMzK7tRAQ0F8WEjzF1+qEi9YVBAuWEZdQ7vWLFohsM+y9bXbOZcFew7n5Vxy1gqcee3nMkFIe+35dNvzqVfU5F5Yq664WI4TirtJ62xDQSHbCMAK01iYdasyvcMNuZFwG+HezV98WQQELln/7tuWiUvtU7AxOCua+un+rqQ8UfXXdYG4dY/skQnAnV2wIgsrcXr8mPzWu22f6b/6tvoETdzNAStgfSY/P1mxzgry3aRonSUdClnSsTEwFEx8NmqFQcBmzz0dWV8Qjb4dN9ctXnL4gBiFftPv7jL3N/d7CFhi2f5DZuDaQ3K+ELX5yoTz7d99QKzGFbGile3dGgk/UjeyPSN0NRMOhSzBFJEQujgt8wiwp7P1uX2ugKn11bfzxqDb6FpUceg6riAVi8UqYcJ3XRa3T9+qi7Pm1FLsv/qAmT37tLiel1/6h/oFTWKHA9K4HYKWX7gkLwC6msmFQpZQxKKwk1gYdVoUcKemxp/wBOymZbWCjYiXLzj5inWH2k1ZD+uYspFYtF+yPcMSa4Pw+sdx9+WLWMh1xT4Gx45I5cb82y+b+YmXRNDqdTk1dha5mrBqmaaRSChkCURqJa01hkIoqQ4ruJII4s+9+Wszd+GUxMBCFljI+qklKgs2nrZghSOHygEE5e28W7tZDz3bbpDz77Wf3VZ8IEDdozcEXVX9DFlq2Aess97tN8g14lrhcuI3XGtNKq5mIVdO05CKElhrJFFQyJKELWRIREWOmFgTdQS54UZOjR8X8YPrNWQtGDfPy49P4bsvHACW1tSZ42bmzDERMexvrWA/YM66uQpqSXusmA3sOSJC1FMRtpCIudeggjZw7UERabicmGCpjez/t7Xvld02Yy1FtXLRSSTy4lgJkBwoZEnBFtb8/GX7mS+nDtThAs28/nNroTwnbuTwu+6RXKyl3S13Dd24llpdEK4rL/9Dw9bWakH+GoRNxa1rcLcZGDtstt70JyZr5+OsNZ1XQYPLmbXCjVjgxK//T13Wmd5TiFlh/orU/lLMkgGFLAk0KGKI91x55Z8ksRTiNWSFIORG+iImn1ZIpiBeNtbkWkrtArG2yee+LVOfDeQP7/u4Gb7h41Xr+Okeel3iqkLMGrDOKGbJhELW6TQoYnAlr5z+sUTYB645JMH88m6Wu2auGwlX8VJFMNbDbWwGaqlN/OpRs+22z1lhK+e6hdxNv0IALRGQUDv54t+LsNeq2aSYJQ8KWYcjolKniM2/+ZyZtoUVcSa4kmhCpLi5W76AQbwunfp23Q2+2w2stPPH/roc87PW2bYDn4utadXrReuEnpE9ZtJamqjZXMnVrBIze1+kMTrpWChkHUwxN7PUnGYFEUMsaPbsr6RwQ8T8nDC1yNSFBDPjx8xbJx6OUiWSBuJ2E08/aqZsDG/0fZ8Tt9N1M/14GqyzEbvOldM/ElcT1CtmxYVp+4Kov50naS0Usg5FM8/rETHEtBbefsn02JgQaiV9fBHLVyyaToiBrQcQtAvH/9pM21rVHYc+H1UK+Faa3IeuQTPye//eupknRMwgUIOBe6bIva/052YW08wz61AoZB2I9tYgeWJ1ihgy2pEEqsTFwi499y0z8czfJsaNbATUsP7WTqPW1YS7qWKWTqeXpZkgTQOJwLPnfiXfa4pZ94CkZJQTkDN1t6AgrYNC1mlIp4AzlUbOtXtT9UUsaIFUCi7cxwsbyAqrBdxNtMXccei/iXXmp5UAfEecDNQjZhn7LPLzl6RJWBoN8tk2s6Pg0+gkpIZyUmal2VGNjP0lEbstssRSMnzb8qx8ZLm/8ff/dVOImDJj3czX7TXP23sE4tptatAf90h71Qhit8n2jmDG5BfKjfRJ50Ah6yDKTWSK5R4sarzxEdhfssQORK6jOylwJV/7uz9NbEB/LeCaX/ven8o98FssuJMrZmqdBamM1FS2mucM6RzoWnYI5QEz5qKBNOLQ2kl0LhjnTurnxDOPWjfrb81m521bM4skYTdNw42bAYgZ2luiAgCxM3RnFEKeTbEgz6o8UDDH0+wEaJF1ApXBMspv/PgGy9L9jhUx9M01eO2hmrukiFWDuBkmxe9WCBNqfJEoixpN3Os4xCqzz6qwOEUXs0OgkHUA6lLWiouVu+A5HiW7+i6k6ypdfPp/U8QCXLRCdtERM+A318K9xT1Grpl07RMCjcxREYNG/HHrkJZCIWsz7hiMteJiaFqDLHwkdCIXSvFjP7TEajPhiJkfM5NKAOsuDl9/j/ShduX0P8XuJ5XpktHZSzIa1KIh7YVC1mYilzIbny+GuBhEbOCag5KSEeoSmiJWP76Y+fcTrjvijxgnAO0z44hczA5tm7qZoJC1kaVayvjhyeDeSHB/2z7Tt3MpAO27lpOnvk0RawCImVub6Vu2MlLTyJgMSRcbL0OXQVEt5qwh7YNC1i7wx29dyrI1Fl9LCZcSMRtN3gzVUiJtADVzpDFwz9DnWlxnjdIJpa3BlI4pCwvBfeDZofayPJo7A//tgkLWJtQay/TEZ+9LW0DrtiBfzO3jXmvc8F3aTf7oLw1ZHed//JdRp5FVsTJjJA1jcOyu8pgAF56L3cdSe8x5Q9oDhawdVKwxfZuHEJfyXNmlRHfPrtvjfr514mubMtl1vcC9e/P4V5bVACtdW94pPYrAxYyrxUTgn1ZZe6GQtQGNp9QM8J8tZ5j3VVzKEIjzoKE0WRtz5582kzZepvhjFqANJmonMQJVHLTK2guFrNXAGrOFIlWj4TFcHcS9EHDOBBqOo4DBkkAvFmR9wL3MXTkbtHzTXYPi3uO5xI1dQKusvVDIWgxETGoqu2unWyDAj26q41yet61LSdYPuI1wMX00XoYa43SmJ7KUQ9Aqax8UshajeWNxsbHyW/+s6bVxsUxg2DaAUY3oUq4/cDGn7X0NjcyERNle+2KBJRyXjlFllZGWQiFrIaV8bsXRrBHgl0Frt+8LujmosbzEfLGmAUsXffT7eWWg96obJR1Du8kOgaH3YJUx27+1UMhaCEb5liTKTLjHBLg3sMgwhJk2Q/JHP7pkYzmspWweSGe5fOo70feqfszsc0PcEvcf8bQQcD9hcTNBtrVQyFqFBPkXoj/0ZYttgZnRmsrdB5YtAzJq+Ms/MKS5TJ76VmSVAVfMIGQgd/lMeGN5UfXIaPAM+rcOClmLkCC/qbgeoeXWGkP3zHApdQBZvyDNnXuG1lgLwLNAky/gZ/3DKusZGZNnFZftn6pY3HHLyfpDIWsRpfzCikF+gORXWT8QH7vEdIuWcdkKWaEyQIs/IhNqMJFXtlDpRtsHQX88a4mJkpZAIWsF1sWAqxEXG5P2km8+J0F+DJYRWg6XktZY64BVhn7/tYIFqJhlB3eJ1ZWbPBO7Pd3L1kIhawH1uJX5uYume2SPfK9q71f5ZGys9Uy9snTP/ZQMDfoXYobVW3IvaZW1AgpZK0BVPEY4CriVKBS5yXGZR3zM/V2tgfz0eclxIq0Fo04hFgb80anQ/hIsVJ6dTyqdLbuXxbwhzYeDj7SAonUxUqnwrUYBQQ2YuJV924JdyrjtAJNCz/A7gr8vXHnDJImZM0/aF8z+ZR0wdtkQAJ7ZItxLxMz8DhplaL5MOeXGDBrSXChkzaZYkDhJqqsvuBhBYwT60ctFKAkTzPzuSZMkuq2I3fS5HwWX/frRe0wuQWIG93L0wH+JAv6uy49QAGKbRVuRExp9HKGEUm7RPn77/Dmgb1Ph3W0ysMaAuBoBNLGyqxIfc9/qKDgLEy9LkiZpD/lKk6TQQC9dlYqZuEoYWGSyfpFpGM2GQtZsKjGSuPiYipQWCr8L67mzjI21G7iX7gtG5zVOVqtHjHKcrGBIc6GQNRkEe+NyxwBqKzP926JaLuAWmtmEuZUbETwDf8QlgGeWsXHNWhZzSoSM7S6bDYWsyZSFLBO7HG9zBPn9pEvZtuJakvayMFGuuQyJWc/w1TLaEuJkIeQlRous6VDImon+AccIGawxkOkbDcZgclbEihxqrO1o87EQqV777GSdmOeU1jgZE2ObCYWsiWhsxA/0R4OILExLIejq375sGa2xziJutKUuhAXsZ352Irg8VSlipRLdy2ZCIWsiJaNv4WqLTN3IwvwEVgpaZJJfdpFC1inkYl4q2sBfrWufVKbyEmNTpaZCIWsmkUUWvs2IraSy3VWBfoUWWWeBFItQHFOeXdpOcT1daJdNjJM1FQpZM8FbOBV/i5H1DddkWVZ4hbjhx0jryU28EqyMARibND83EbstXmRL1jlpBhSyJlJCRn8qE/i9XAAKVqgwq296P6s/R4usY/BTLFxRy/QM2lrLWo3DbTErlgxpHhSyplLbnSguTJlsYKRxCFtcrwqkfWgG/3L3smeF2mW8qOhaNhMKWbNJp2ouTlWCxb5rmZ85b0jnEQoDIOAfsqyVWnmEZH2gkDWRcu5Q+BbjDY4/dy0SfrJlkRZZx5GfWh7wF1ETEUO8/8rSbz6stWwqFLI2IW9wI2UgmH+0JHGkkwgG/Pmo2g6FrI2gDLiD8HpLDek83G581DpDv2SV1pfyf/jFRJoJhaytVP78U3ylJ4LU8hBA9Owcw4zPs/VQyNqE/LGjdjI3tTzmQjqTymPyxaxkn2E5VEZLrF1QyNqE/NF7f/ehGi/Sefgvm/JzMybUuy9pDRSyJlKre2ONjbm1lmln/czALkM6C21XuQyJlyExdrjy1bOqpUKARa2Z8O42m0AbO/cP3U185Zu8s8F4lqFnhFQZ/TkUGigx9aLpcPCRppKp2cYOtV3FQGNj6Q9+aLdJKsinGv/hX8UsS25+HJ6X1lYqmMfYlWhvGfUcGxCzUFM1sn5QyJpIaoX+2tOZLumTrLxuqqogYB6jjidx4BGI1cVTf2c2Et2jN0S1lG5tpTyzYk7cTl/klrB/A2kWtWZC17KZIC5Sw61AX/0FqfEK11R2j+41pDNQoXLR74szb0ddMYVcTxkOjkWtqfDuNpMa3RyjEKTtH3+psBjs7x3Ls4PJdS83Gt1b9wZrK01xUcYmzfaNhjfUFxnbWzYVClkTievmWNMs0n3bjPQjW+k5wXdZNoJFlrU1edlKbV6ScV1LBfOLs2/LfHZge3C7qNcL1lo2FTruTSSum2MJEFsrDZ0qgtzUOdNXGUnJBVZAUukZfofZ85G/MUPX/L58n3zlcfP6E//DLCRolHEXCBnwxawwW+7iOi41Ixq3gcH+psLXRDOp0c0xcsYylb7I3J5g3RgLCk9s7lKHc81dRyMRAyN77xZhSyJ4BnHWsfYMm+kdDcc6YwagIesLhazJ4A+4WAy7lggQZ/u3V9VM+jVfagkkCQj0yN5/s+x3CFu3tdSShj4Dv8E4KFghwzWF+vMHGNeU8bHmQyFrMvImrlFziSRLDEKCgLGfggH6r/1DkzQyNWJiScwj67/2zmVpF4J9QeVtjCzjBPqXVQjYddI1Rpon6wOFrNmIkJWkdlJx/9i7KjWT6EbZzVNS+mwhSho5Gwebev3/L/sdv2nng0mid9f75NMXqdyVs/LZFVO7XFJLnG5l06GQNZl0lF+Uj35zhUoz+BcrvY8C9xPLYbUljfF//KsqMcM8fksaSIGBaxlKds1NnpF2lpmY51NifKxl8A43GwT8keGPUXayfeWfHKsrne2RN/riZVsorjm4bDkYfNdHzeSzXzdJAlbZS9/5jLiZGGAlqbWV/RWLONT8aHH6rOm2z07yAQN9+ZeQH2ifPYWs+dAiawHlP/R8VaysKqA/skdiR4ten/BRnGzPYZNU4EomVcTA0O/98TJLGSCuWVycMV322emy5QmzhaUUHNJUKGQtIKVxMicNw+27qmfrdTLvxsnc4DLyyXp2vs+Q1oLYmLr+vpjpKPDdW/ZUPS9F4mMY1zTdbUjzoZC1AIx7CEqBni5kedeAxGLm3nwuWMUPBt/1EUNay0DMPcdzWZgcl2ZJmufnu5alSrOzNIWsJVDIWgFEyVbBh7rs0T/+3q1j1g9bFPcS+AWj/51/GI2BSZoPAviITQJ9qUSNxO0zKuWmTd+Om2Mb/KNrHzxzk2YRawW8yy0iZYP6bhqG76r0bNsnCbLzb52q3q5SgPDmH7bxGtIaRm75bGy3POpWZod2xSTBVtzKbI8hrYFC1iLSMe5lFNi3Itaz7QazcGlcCkKoyxgIGa2y5gM3v2/XAZn3nwOak81ffMn0bN9n0l2D4R5h6Va2HApZq3Ddy0rtpWuVSdB/ZEzm551YmQua/gy/+z8a0lwQG0PuXqi2cu78MzLOQv/uAzEjipfKbmW2m25lC+GdbiHprv7KH3r5je3HwVB4uqy7AiFDH2VubEbFbvg9/yk2AZOsHVhjW977mWW/S48l1hrLTZ2VdWCNuc9PP/VFpRU8pDVQyFpIKtO1lBxrlndvjQkuDd7otWJlo7f/uSHNASLm9zmmLFx8WYL8A1ffFi3zY2TFxVl5xmkKWUuhkLWYdLZXYmBu20sQ5YwNX226h642cxeeixqS+ykZqMFkXtn6g3atg3s/uqyWEiBhGUH+LvtsXLfTRYP8YnmTlkIhazFpNFOyb2x5cztoZ4sA8RcMaAEXM47tf3CUgf91ZvT99y77TQVt4a3npSff/t3vC64DirlZiYUyyN96KGStBn/ojlXmW1wAb3zEYWbPPRN1uuhbZVg+8t7PGrI+bLnls1HjfN8iwzOYe+s5SZHRvseWYS0x6bIHLiWD/C2Hd7wNxFllQAvJ4J7D0qB8+syxYO0ZGH7PH1sX81ZD1kbfNXdW5Y359xk1lbKetca0U0wlErvKs0xn6Va2AwpZO/CsMheNlWV7h03vjhvN4tR5s4DuYky5e2y/sG3/g//O0ZbWAGqAt/7+UuWJb/kiLoa8sX5bCYNBVGTQGM/iwnMs5uflmdIaaw+8620iZJVVtdWzBaZ/581W0IbMzJnj5WHHvF4x8JkZ2Gm23/UVQ1bHVUe+YrID1S6lgjFHZ889bTLdQ6b/6gPLRCzq7hq93qKmktZY26CQtQvHKsPb3Bcx+Z7usi7mEbtOzrqYx6vW0QKlvWOM3P5nhjTGVnvPZICXiqWrqKDNIUa5MG227Pt49LvvWuLZSU0lrbG2wjvfRqSaXqyymXKwWDtbtAVC5xGA7ttxk/RGqr1j+O6PNl8KJXKSMMP2Xvl9jblihhpj5I2hljJ2lHH7zKK8MaZctBUKWZuB2yLZ/otzUWFRF0a/D1xz0HQNbBcLoVAZfgz4hRA1bxSzlcF90uA+8FtQoJZy5vWnTMa69QPvuH2ZyC0F+OeYN9YhUMjaDLL90WC8mJ+rSsfw42FD139Q2vhNvfrjqkbl/ifFrDbu/ZGWEo6LDiBiV17+gcQmh/d+LJxqAWCN2WcGl1LcStJWKGQdgFhl1j1BcDk0KrkE9XuGzNC77pGCNvXqj4IukVpxFLMwuCfaBCkUtMcnXhS4xxiGDzXH7shWkajZ+fz8JAP8HQSFrBOAUHUNLsVcop+rCxCaLw1ee9Dkp86bqfFjwWRanaeYVYOY2HCgHaX7feb1E6Y4PyEtK3Cv1SpGiwv3PsszUpeSAf6OgCMjdAjo9iVd7BN3BbWVqcrIPLLMcTV7beC/mM9JWsCctdL6dr1vWWxNC54UXBuonvzl/zKbGdRODlU6pQzFxQBaUcy9ecoM2OA+Ui1C7j1ALSVdys6Dr5MOYqkWc9qkHEsMwuRW+6OgoRNGBP+lf6xUdb9mrquJ2sydH390U3b9g2ve+bFHgyLmgns4f+EZ02vvaX+lZwtdr2pdsZhnKrWUA4Z0DhSyTkIy+kcwY/ILk+V+rWIsCDRhQtu/OVhmFTFzM//dGFDP6A1mxz0PJ3LU8tWCpls7PviwCD5w7w2IcsXsvYM1hns0NHYkWqYvjihnTONilmzPiDwr0jlQyDoNzUmyIlaoNBgPgYI5NHbYdI/uK7uZ58rtAV030xU/DAK8/chXxN3cyMCVhiu580Nfi4Zyc2sm3Uks2nNlS8wVMcXdrlipiBFLjHGxjoMxsg5EYi+VwH8xN2PS3QPBeA3mIWZly+Jp6b+s/5o7omWhgTFQAYCunCd/8T/N3Os/NRuJnh23mtEPHK3qxULx56fPPCntKCFiw9fdtexe+cF9dHaJFwzjYp0JhaxDKVtlhXLwH66i/e4O6usWNIhZpmfQzJ592uSmz5lh5JxZy8G1znRdINbZXX9jpk//P3Pl1980henzJslIw+/b/kxc51B+netSornX1KuPmby95gFbO6k9Wui6riULRMTsJMF9Jr52LBSyDibdPWQL3lJKhhYkLXBuUybUXpqSETfzyks/MMP7PhYNHutuA7SwYtxGTJf/5etm5tUfJk7Q4EZiMJZBO+Fa43LDdMrPXiznidmA/YC1XNH0y7XUYkXM1iCn2YllR0Mh63AyPcOmsHA5VszcT1gXWRsXmh4/Zi49923rZh60hfXGsnvk5EL5bHnvZ83A9R81C2+etKL2jY4XtCUB+w8i9vJbwI10Y1wYA0HjiCP7Pi7xs5AlpqiImVQ2OgbpXChknY5k9W9ZJmahzv1Al3WzYI3BKpt74ylTmLso1hosCjfPTNECjIKN2FL/dR+21tk/mtnTPxRh6yRQE4n4HrLuU5X0B1+09LfIlbSxrenfPSmN7rsGtpmh6++RVhLutu58FBOriFi2dwtrKBMAhSwJBMQshf7MKvgxMDR5Gr35T8zM2V9JasHi1DnJVkcqgm+ZuakGKgDqcuZtvG36hf9rZl//SdusNPQV1n/9hyX+hS539JzdT8V3J3OT4xLUR1ysb+fN4k7GVQDod4pYMqGQJQUVs9xUOW5jzLLgsytG0jFjRbxgnc1Yq2TRCpNrnYVcTfd3jBi09f1/Ln2d5dBT6oWTZv61n1qBeEWGRWvKZdpzQ/9q6H667513Rp0elm9BquanzkvvFVbAFmfOiagP7PmgjEzlp1b4oo4aYlSuYCBluPQUseRAIUsSImbD5QJnxaxUzNuCiiB0OLMfn+ieeetNn5JKAEzoY6vnqhtN/86bIvfM39atFNBlsIa6rMAMV7LkcxOvmNyll03+0mmZR2uExUuvmEaASHWN7rWfO03XyF6ZxzFqWVpKKLCPXnQxjB5iYalsj62VvE26C8fYB7peqOYXya4le/6SYmHXZUwseVDIEgjyykzFDUK2uWSaV2ow/eZKCqyz3m3l5FkEvhcvnymPCjT6LrFaXBFzrZXQPsvCtlcm4AoerKH8zHlTRPfPAbSpFFJAQlahnwLhr+PGv6L1Cwv2mp4382+eEjcS8T4kuCIWFsqlq74/RXHZtRE4UyySCYUsoSy1y7RitnBJMs5TzujWrpBFsTNbsKVpk7WupDXAeVhoL4mo9Hk9oaY9YfQJuWkyD9ewe69ZCT+epcIZEipf1JYC+Qtm7s3nRZgR1IeAac8VJnCefqoFtoElWT7vochyI8mDQpZgJEkz0y1WGSyhdLZQrgTwrDE/MRYFHYUeMTMk0SLDHRPiaRC5uMx4V9h8S8fNaQsREkXfPQyJpy+Y+L5oKx7mrQuZs1YlvOqyEFsBs9fkn4cvXpUZU7KxsHKycTqyaElyoZAlHRTE3q3l2jYUTGuloFJA2mxWCmdo9CUAAcDAGhBBcTmtdYYYmgTct7yzLGr926J9hFw9d5nu260F1eXpGKHwz8nvsVUpzFwwuStnzVzF+oL11GfjfN1b9ogo+wLrC9hSPCxfHvWoMmCItJ1kUD/xUMg2ArBmEDdLZ0TQCvOXysPNoV1gqlrMdL5KQKxwweWEe4lOGzHIyYKNOWGCqEHwukf2SI4aLEDdLiSQfmP1ldzTZe6iCqYVZIzpCesLlhfEFmtAtHAuvdv32R10BS3B0PnIOpEVhvs1bO9RtyEbAwrZBkJdTdc6QywNsTPfYvEtLHwi6J/ZNiQikZ+/YisEfieJpKiNRPoFtsz2jZpM37bK56h8pmrEluKEpspisxYWevqAAKMZEVzeojPICtJAenbdbF3id5q0k8waV1mgRNdq70N5pKpSpbnREK2wDQaFbKMBlxIpGelspfcMG8xOzZYD+eh5NibG5f+O/uqzvTdJ+kIpb60jWxOJxFoIDdIucpdEF8p6kO42XXBB7f7TmZ6oXaJ1MEU0JJjv1GLinIoQl9yM5MVByKLzwLHtvnqs2whLULriSXdVxc/cYfP8869KscCgxotz8on7krHXhHtANh4Usg2KdsVcts7mrZBcEctMLDTrgkbrVQQiFFvS76muXolFYQJYFxZT3opQYdZ+zl2UHKyitagWHcFSqUy585UvsKxgHfVaNxEWXbZ/uxWaEbEK0/Z4fjtI3x2uVRtpL7iSZ7dYdiOZVrHhoZBtcLQPLRW0grWExDKzMTQMRRdXwwf8pkvueumB7SZrJ7P1uqr1dXnBuqaiWm5szpjgYLfu9qFKAT/m5q8ftsAqAlapxSUbGwrZZqDibqJglzsJXDClnBUag9/7ZaCTUK2jfvcD8W6jc13H/QR+LMtft1YtqH9cXT+0XM4FbqqdYIlRwDYnFLLNhApaqb/sCkYxNCsKNs4llQLZpVpJxa+J1HyuUJqF4rqsuq7/m7uef8zQuq4o4vwRW0P8Tmw9CtimhkK2GZGBZcsxtFLeClphvlyzh1jaYqbievZEgfEq961UWiYw+ntI0HyLTn9zt3PX1VYCbqa/nAOsLxEuOxUXIpcV5yhWJYP4mxoK2SYHFlgGVhjGCKhYOBA1TGLZoGNB5I6lM5FY+PEy180LJeGGWgX4VpqibqssR6wLgXvUahYLRqsMRGi7uu0p9dL6IgKFjJSpWGlGBz5R101qAGeW1pMUDohaWlI8DObt5Me8XNHyrTe/KVLJHk/iW6WCdO0tAmaKSxUF2B9qXDPZcntSihfxoJCR5biiZqAneasxeREbCFupgNG2nQGEK9uIoFXiVSmDFI/qAD2+SfaFFSwMQCzrloqBw2dRWyBpIuVkXraDJLWhkJEVSVn3MpWt/lOBuBkZAR1WVKEsSLCy5BOzizE7S5d7TxNLLlX+ni5bdbDwKFpkNVDIyKqAuMHoopNHOgG+/gghiYdCRghJPBQyQkjioZARQhIPhYwQkngoZISQxEMhI4QkHgoZISTxUMgIIYmHQkYISTwUMkJI4qGQEUISD4WMEJJ4KGSEkMRDISOEJB4KGSEk8VDICCGJh0JGCEk8FDJCSOKhkBFCEg+FjBCSeChkhJDEQyEjhCQeChkhJPFQyAghiYdCRghJPBQyQkjioZARQhIPhYwQkngoZISQxEMhI4QkHgoZISTxUMgIIYmHQkYISTwUMkJI4qGQEUISD4WMEJJ4KGSEkMSTNYS0kd/+9rfm/PnzZtu2bWbfvn2GkNVAIVslFy9elAKIz7m5ObO4uGi6urrM8PCwTNddd53p7++vuY+f//znNZdjf7t27TKjo6Mr7qsV59sMLl++LOfU19dnWsHrr79uXnvtNbnmG2+8se71ca9uv/12QzoTClmDQABeeuklsSRCy1AoMWH5NddcY2666SaTzYZvM9ZbCYgPgLWyGoulkfOFmNVTuJPM7OxsdM0Qp5Xuqa7fDJGHSGL/sEYxkdVDIWsA/NGdOHFCPgH++GAx4RNiBUsHE4QD6+APdWJiwhw8eLBmQYDgXXvttVW/qchcuHBB9oV9ouBBbJp1vhAzfN8slgeuG5YZ7kk7gKWHZwwxpZCtDQpZA7iiAMvFFxUVKwgTCokKxLPPPmsOHToUu19sF/pDRgHDHzm2h2V26tQpEbw4C8/nl7/8Zd3nCxHD9J73vMdsJnBvIWbtcKvJ+kEhqxN1A0BIFHwgQLCgID6u69Yo2Mctt9wSuZh4i9ezH5zvlStX6j5fLI9bB9Yhzh/7wwQhRcGH0EIEa4FtcS64B7othAOCXK8Vgm1x3Xo9uCfYfi2xQ+yjVCrJ+eEFdfjw4bpfED6ha6x1jhob1evRbRU8bwprY1DI6gTWFUDBrVeQsJ4G2PHHuhohAxqUxx8+guPNOt8QEJDnn39eCqsLzgXXhuPEuc5Yx7UK3d/ruR84JrYPxRI1bgULcjWuIe4pttXz+8UvflHTao4j7hprnaN/Pdg2tD2pHwpZHeAPT//Q/FjWSqCwYnsIEPaxFguiXvRYoNHzdYHYwPUCENLrr79ePiEwKMCw0jQO54sZfoflobWjuA9amLFtXAWEC7bHurq9Wn8a18N9hYjAmsJ5NYq67rovfDZSoRJ3jfjuxh5xjhBJtUDvvvtu+cTvuD7fGqY11jgUsjrAH6XSaFDWLWAI/K/mj1RrzsCWLVtWXN8933rWjzumWnUonH4FAO7Dzp07o7ihHwfE7yjQuF5f5HBPsK0KVQh1jUPbYx7HgtsOMcTnaqwpAOHC/YLlietFGki94l/rGoF7jbg/6r7qevpywifFa20ws78O1mL2r8cfqGu5oHCsRD6fj+ZXG/eB6Op1x1UA4NoQzwGu1erOx8V7VsrLUhGF0MTdQyzD9Wk6xWrB9ekx4EbX87wbvUasq3FOsv7QIquDtYiRG1uKcw+xjl94IEZ4k2sVPahVqOPAflYjZufOnZNPWGO1jqmpHDiOWpwax8P11rJgtbbWFyH/ftQSqe7u7uherTaFAecJi+r48eN1B/8buUaNb8ZZn2TtUMjqwM06R6FqpMC4f7xxBUNTH2qB+FS98ZuhoaFoHgVuNQVcrbp6RFDFRMVHt60nbhXK6HfF/+TJk6Ye/MqIRoHgvP/97xdXsJ7gfyPXqEKWy+UMaQ4UsjrAH6JaHdousF5gUYG4XDGAt7pvraFgauF0A8X1gLiYni+Ovxoha8SK0wLqX4Mbq4tjJQGC2wehXInVBPt9tL2nBv8Re1upkqWea6xnHbI2KGR1gD9mWERa04ZgcD0FB6KHoDWolW+FGivf2kIBf/LJJ6Oge6PBbD1fHL+RnC0FYohWBZhqAYvPt070WDh3WCK17lXIbYToqxCnUqkVc9XWEzf4j2cdd+7uNa7kvqsbun37dkOaA4P9deI2qtZq81qggGrqArZrNJfLDRRrQu1qzxfnUc/5wq1SUdJKBW2rGcfLL78sn67FqRYsqLUtlrkVEy4QYqApHnE0I+4EK1AFLG7/7jXWcn/da0RybIi4e0DqJ4P/7rzzzpFSqfR5zPf29po77rjDkGoymYwU1DfeeMMsLCyYM2fOSAHzq841Hwm1X8ViMaqax3310UIe12i4p6dH9v/WW2+ZyclJc/XVV9edT7aa84UlMj09LcfB+ULEcFy1mhDP0uOrUOPcAGruBgcHo2NjwjLsD2IwMjISbYv9vvDCC+b06dPRefhtHvEdFQ44Z1iFmhSsYJ+/+c1vZD87duyQe1UPboPxuJcLzv2qq64SqwzPEPjr+9eI88T5xV0jLD0/cVdbAuD54MWBbbEdvjeSN7iZeeqpp8z8/Dxdy0bAHypqszR3Cm6buo4hICRrbW7itg6AJfiBD3yg7vjVas7XTbXQnjBgFWnbUVyLG7/T9fxCivNWaw7nj0ljgWphYV+wUkLnpDWJeu6wejD5x9f9rUeMzMUN/sfhXqPeW/8aQVzPJXD5tekbwgh4UeBlgmtZbV7cZoUWWYPom1kLVCiQC0HYv3+/FPBab9aVLDJ3f7AOcKxCoSAWyHqdL5Zv3brV3HrrrVH7UBccy90Wn7BSsB6WHThwILaJEM4bhRIWBrbFdpoFv2fPHtl2ZmZGRDrUCwXWU0tFj+0eH/vAiwLnXy/1WGQKrlst4rj1cY2I4ak1q9eoy3Bf4xJsNTUDVi/uESZYebivsAjJyqhFlsKXo0ePjtkCIkEYuAD33nuvIfUD9wDWgsZTQm9gTaBcS5Oh9cINUDdqLbrW1GqOu9ptgWuJdWomvF5jqCa6FnptzPBvjEceeUReBHQt1wF1BbS7Ha2+VwtDXUPt8bXd8Y+1FJZ2bQsaFYd2sJa2tIyLrR4K2TqhtYya3Oo3m9HmPPxjJWT9oZCtM9qTAawwdTU1VkQRI6Q5UMiaBFzIdnWhTMhmgwmxhJDEQyEjhCQeChkhJPFQyAghiYdCRghJPBQyQkjioZARQhIP88gIWSe0pwx0FrDevXGQ2lDI2oCO3K39UWlDaPzxNzoKN+kctEnaWscPII1DIWsx2q9XCB1pB31Uoavp2267jb0h1IE2B8O9amW32KRzoJC1CLyl3QFptX8rtcJ0FCLt5x/9vOso2qQ2es+0bzCy+aCQtQiMyKMihv7oQ4PeaueC6MsMIoZO+QghK0MhawFuF9Nx3R67wEVayRJDPEb7OYM1h26SIYTuoCM+GozW7oQQp1N3FttjgsiuFKiG5agxvkaPjUA41tdjw1LVbp0xr9Yolq20b3056AsCn27X1LgWf2R2PYbGJ9G5pL5AVmrkv5p7TloDhawFuGNb1jvIbi1CcTaIi462FCeWGoxGAUYnkG5QWns2RSGvJbbrcWxYm35vsa+++qqMyOQHymvtG/ty+3zDtu53vzdeiJB/3bofdU3RT39oTIRa141tQxY2aR0UsibjFq71FjG1OHTQCggmChWW1+qTXoepw/loP2nYHtvpOJqhcQTW49g6dBr2DQtIRxqHhYN7hd8hQBiUBL9BZOLOCx1ZYhtYZhhpSQd7Udz+33TwFj02rh3Hdo+BdUIjjLvX7Z6fe916T0l7oJA1GXdcxLXmFmlhBnDRXLHQcSVR24mCjfVQ4ELWBQo/XFf/fCBMcYMC13NsDEALoap17ND2AN9D4olzxHk99thjIjoqWEC7h3YFK87FU6GBeOp4oe4xIEzHjx+PLCytNHCv27cI9brxG7bl+JTtg5n9LaTeYdy08PiujH5HYYyzeFQQIFbq0vqgkIZEFYIwNjYm8/7AtK5FEnds7FePDVcxRNy549hxuXNYpm6ijtrdCDrwC4hzASFKOiiwe9/0umuFBdxtSXugRdaBqJsH/JgQgEvkxoJ81F2LK/S7d++O3VYtGj+OpMPIQQDrOXbcCN0rJfrqsHP+9qFh9+pFzxfWKvYTty+9dvf69DxWqgiAiNcaVZ00FwpZk9GCDSYmJtZUu6Wuiw5wshJxBbaWZRg3roDuq95jx2W3x7nXGuuqNYDwalFrDMJea8Bdfxs8K73nK1nTrLVsLxSyJqODsOLNDpdlPRI2sQ/Eo1aiXle2Eeodi6DRY+tQeu6Awi46WvlqUHHGc6jXBWxUmNgsqb1QyFoA4juwNuCyQMxWO0jv0NBQNLhuqzPYEQzHsVFg1/vYsJRUpFDBELLatNXDaoBLCWBVQoQbEVm97lruNEAlBGkfDPa3AG2KBJ5//vlVu09qTWhSZhzqSq0nKr7+eJ3rcWy3ti/OEqpnv3G1hqj1hHhBhF988cXY7fUl4eJed1zlCfbL+Fh7oZC1CFT5o5Dijx4pCphQOPykVMSf4nKS3PQExHo0v0rRAvX444+ve8Gq99hI32j02K4FFqrtXCl2ptvjfNxzctu17t+/X+Zxf3Hv3coEzfU7ceKE5JG5uNeN54Jrc58ZtsP9aMbLg9QPXcsWARE7ePCgFAbNVapVOFH4QtX9EERtfK4pGmrFNLsw4dhIKsX5r+ex9Vp1n7gvcOkAXDYIh8YZQ8BqUoGBiEN4dN0Pf/jD8gmrWMVW773moOF3FSe13NxKDyTZQuTctBh9Kel2sJZhKVPQ2gOFrIXgjx8xIBQiuCkhF027okHBC9Ug4jcks2IfsF5QYLXwYBkKPEShGf2ZYf/++bsFV5NDV3NsbIdrV0vPbcKEZTi2tgoInRfuCUTWbbaE84CrqDEx7EfTJFQg3b7gtMG+D87h7rvvjlou+Ofnji5P2kMK/x09enSsUChInfrIyIi59957DWkNKBAIQiNNw89Sb2QfoB0pAM04tgrMau5HI9u6L4BGjrOW8yPryyOPPGImJydpkbUbCMBaRaCdOUzNOPZaBKKRbVd77hSwzoPBfkJI4qGQEUISD4WMEJJ4KGSEkMRDISOEJB4KGSEk8VDICCGJh0JGCEk8FDJCSOKhkBFCEg+FjBCSeChkhJDEQyEjhCQeChkhJPFQyAghiYdCRghJPBQyQkjioZARQhIPhYwQkngoZISQxEMhI4QkHgoZISTxUMgIIYmHQkYISTwUMkJI4qGQEUISD4WMEJJ4KGSEkMQjQjY/Pz+pP9h5QwghSWBysixdImQPP/wwvskvEDJdSAghncr58+ejede1PKkzL774oiGEkE7mwoUL0XwkZKVS6fs6TyEjhHQ6J09GtteSkOXz+W/o/Pj4uEyEENKJIPx15syZ6HskZJU42RP6/fjx44YQQjqRY8eOuV+/UZV+Yd3LL+s8LLKnnnrKEEJIJ3HixAnz7LPPRt8zmcyXM+4KP/vZz8YPHTq0NZVKHcT306dPm/3795vBwUFDCCHtBjWV3/nOd9yfvvbggw9+K+OveMcdd5ywCvcRO7sL30+dOmW2b98uEyGEtAsE97/73e8ini/frcE1/tBDD30U88uEzJpt84cPH/6hnf2knUawEcQM7Nq1y2SzWUMIIa0Cua2PPfaYefzxx6tELJ1O/9FPfvITSXpNxW189OjRsWKx+M82bjamv42MjJgjR46YPXv2yDwhhDQLCBji9JjcFkcqYg888MB49FutHUHMrALebzf8z/6ysbExs3PnTtPb20tRI4SsC9qyCLEwN73C4YmBgYF/d//991c1P0qZOrjvvvs+Y8XsS651RgghrQJWGLIqbEzsG8HlpgEgaPYD1tldhhBCms8TdvqmtcK+51thLv8KHdJLJwWJ6/wAAAAASUVORK5CYII="
+    static let fallbackQrCode = "iVBORw0KGgoAAAANSUhEUgAAATIAAAE0CAYAAACvn7/YAAAACXBIWXMAACE4AAAhOAFFljFgAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAADaoSURBVHgB7Z1pkFxXledvLrUvKpVkLcZGZSNLNLaxkU1bot2S24NZJwJm6aEjiOiBGD5NdLuZIEwrpj9goukYh80H45lv4w7g0wARQ0NEN9M02G3JLPIAttxYxptMCdtabKtUUu1ZufT9n8zz6uat+7IyqyqXV/X/hZ7yZb79vbr/d865595rDCGEJJyUWSNHjx4dM4QQsk488MAD46ZBGhYyK1x3FYvFT9jZT5ZKpTFDCCHrz0lMqVTq+w8++OD3Vlq5LiG7//77R2ZnZ//CCtfn7dcRQwghrWPcTl9+6KGHvhG3wopCBgusUCh83c6OhZaPjFDXCCHrx+TkZNyicWtMffarX/3qE/6CmkL2xS9+8Ut2w/vd33p7e80dd9xh9uzZY3bt2iXfCSFkvZifnzfnz583L774onnhhRfM5cuXq5Zbd/N+625+ueq3uJ35IgbBOnLkiIgYIYS0ipMnT5pjx45VCZovZpnQhr6Iwfr69Kc/bfbu3WsIIaSVwPN797vfLVaaI2Z33XnnneanP/3pMXxZZpFVYmL/rN9vueUW84lPfMIQQki7+f73v2+effbZ6Ls1uP4IMbNlFtmhQ4cgYhLBhyX2qU99yhBCSCcAy2x8fDyyzKyLedeHPvShb1YJ2X333fcZ+4HJbNmyRUSMwXxCSCcxNjYmVlk+n8fXEfu5kPbW+ZLOILDP1ApCSKcBXTp8+HD03bqXfxEJ2Re+8IVPmkquGKyxW2+91RBCSCdy8OBB11sciYQsnU5HEX34oYQQ0sm4qWCuaxmZYPv37zeEENLJoDJSCQoZ8jYIIaSTcXVKhMztigd+J2sqCSGdDnRKKyTToYWEEJIk0oYQQhIOhYwQkngoZISQxEMhI4QkHgoZISTxUMgIIYmHQkYISTwUMkJI4qGQEUISD4WMEJJ4KGSEkMRDISOEJB4KGSEk8VDICCGJh0JGCEk8FDJCSOKhkBFCEg+FjBCSeChkhJDEQyEjhCQeChkhJPFQyAghiYdCRghJPBQyQkjioZARQhIPhYwQkngoZISQxJM1hKyCUqloTDFvPwv2s1j+LJXsVFhaHiCVqrw7Uxl8sf/sZzpd/rTLUmn+SZLG4V8NWRGIUqmwYEpFK1IFiFd+2ToiUDJBmDImFbuzYlnwgBXCYikX2FfW7sb+aVpRS1UmQmrBvxCyHCs0EK4iLC77GQlPxYJKZ/qWrCgITWptEQqx3qwlV7LHE+sOApefX1qhYqmlMt32sN1ybEJcKGSkjBWrYmHelPI5KySL5d9EuKyAdHVXLKMu0wzUmvP3j/OAqOk5lQo5A4cV66WyPRQ1EkEh2+RAIIq52SrxSmf7rFB0i4jhe7sQwYK42fPRcy3lraVoBa2Umy6LGqy0bK98ks0LhWwzAusrPyeTuI2ueDXJ6loPRNi6u6SqfUnUFkxhIScWXbqrn1baJoVCtpnwBEyEwbqN6UxvWy2v1RCJWmlAxKy4OGstS2ulQZQz1u3M9lPQNhEUss1AQMDSPf1Ntb6KC1PB39M9Q2ZdEWuyVyZ1k1FRgKn8OwVtM0Ah2+CIpbLOAoage2HhisnPTZjC7EWp2cQ8YldxAuYDQct0D0psK9u3zWT7t5m0/Z7t325WC64r07tFUjzK1z0v5ySCZt1OsnGhkG1QYJ0UICq2UK9VwCBci1Nn7XTOLE6fM3krXi6oQYQoZftGTXpwlxWkJatLLTBX4Iq5qXLAHvudPm9yk2eq9tc1tFvEDZ9dQ1c3HshHvAwiac8L90BFTWJoVtTIxoNCttGAG1kRCnG7bIFeTeGF8CxcPmNFZlwETIHl1LfjJpOxnxCbTA+sqh6zFsSim52wVt1FsfDwOffmczIBCFr3yJjpsROup14g3BBXscwqMTQk4dLd3HhQyDYQxbx17RanRMzK7tRAQ0F8WEjzF1+qEi9YVBAuWEZdQ7vWLFohsM+y9bXbOZcFew7n5Vxy1gqcee3nMkFIe+35dNvzqVfU5F5Yq664WI4TirtJ62xDQSHbCMAK01iYdasyvcMNuZFwG+HezV98WQQELln/7tuWiUvtU7AxOCua+un+rqQ8UfXXdYG4dY/skQnAnV2wIgsrcXr8mPzWu22f6b/6tvoETdzNAStgfSY/P1mxzgry3aRonSUdClnSsTEwFEx8NmqFQcBmzz0dWV8Qjb4dN9ctXnL4gBiFftPv7jL3N/d7CFhi2f5DZuDaQ3K+ELX5yoTz7d99QKzGFbGile3dGgk/UjeyPSN0NRMOhSzBFJEQujgt8wiwp7P1uX2ugKn11bfzxqDb6FpUceg6riAVi8UqYcJ3XRa3T9+qi7Pm1FLsv/qAmT37tLiel1/6h/oFTWKHA9K4HYKWX7gkLwC6msmFQpZQxKKwk1gYdVoUcKemxp/wBOymZbWCjYiXLzj5inWH2k1ZD+uYspFYtF+yPcMSa4Pw+sdx9+WLWMh1xT4Gx45I5cb82y+b+YmXRNDqdTk1dha5mrBqmaaRSChkCURqJa01hkIoqQ4ruJII4s+9+Wszd+GUxMBCFljI+qklKgs2nrZghSOHygEE5e28W7tZDz3bbpDz77Wf3VZ8IEDdozcEXVX9DFlq2Aess97tN8g14lrhcuI3XGtNKq5mIVdO05CKElhrJFFQyJKELWRIREWOmFgTdQS54UZOjR8X8YPrNWQtGDfPy49P4bsvHACW1tSZ42bmzDERMexvrWA/YM66uQpqSXusmA3sOSJC1FMRtpCIudeggjZw7UERabicmGCpjez/t7Xvld02Yy1FtXLRSSTy4lgJkBwoZEnBFtb8/GX7mS+nDtThAs28/nNroTwnbuTwu+6RXKyl3S13Dd24llpdEK4rL/9Dw9bWakH+GoRNxa1rcLcZGDtstt70JyZr5+OsNZ1XQYPLmbXCjVjgxK//T13Wmd5TiFlh/orU/lLMkgGFLAk0KGKI91x55Z8ksRTiNWSFIORG+iImn1ZIpiBeNtbkWkrtArG2yee+LVOfDeQP7/u4Gb7h41Xr+Okeel3iqkLMGrDOKGbJhELW6TQoYnAlr5z+sUTYB645JMH88m6Wu2auGwlX8VJFMNbDbWwGaqlN/OpRs+22z1lhK+e6hdxNv0IALRGQUDv54t+LsNeq2aSYJQ8KWYcjolKniM2/+ZyZtoUVcSa4kmhCpLi5W76AQbwunfp23Q2+2w2stPPH/roc87PW2bYDn4utadXrReuEnpE9ZtJamqjZXMnVrBIze1+kMTrpWChkHUwxN7PUnGYFEUMsaPbsr6RwQ8T8nDC1yNSFBDPjx8xbJx6OUiWSBuJ2E08/aqZsDG/0fZ8Tt9N1M/14GqyzEbvOldM/ElcT1CtmxYVp+4Kov50naS0Usg5FM8/rETHEtBbefsn02JgQaiV9fBHLVyyaToiBrQcQtAvH/9pM21rVHYc+H1UK+Faa3IeuQTPye//eupknRMwgUIOBe6bIva/052YW08wz61AoZB2I9tYgeWJ1ihgy2pEEqsTFwi499y0z8czfJsaNbATUsP7WTqPW1YS7qWKWTqeXpZkgTQOJwLPnfiXfa4pZ94CkZJQTkDN1t6AgrYNC1mlIp4AzlUbOtXtT9UUsaIFUCi7cxwsbyAqrBdxNtMXccei/iXXmp5UAfEecDNQjZhn7LPLzl6RJWBoN8tk2s6Pg0+gkpIZyUmal2VGNjP0lEbstssRSMnzb8qx8ZLm/8ff/dVOImDJj3czX7TXP23sE4tptatAf90h71Qhit8n2jmDG5BfKjfRJ50Ah6yDKTWSK5R4sarzxEdhfssQORK6jOylwJV/7uz9NbEB/LeCaX/ven8o98FssuJMrZmqdBamM1FS2mucM6RzoWnYI5QEz5qKBNOLQ2kl0LhjnTurnxDOPWjfrb81m521bM4skYTdNw42bAYgZ2luiAgCxM3RnFEKeTbEgz6o8UDDH0+wEaJF1ApXBMspv/PgGy9L9jhUx9M01eO2hmrukiFWDuBkmxe9WCBNqfJEoixpN3Os4xCqzz6qwOEUXs0OgkHUA6lLWiouVu+A5HiW7+i6k6ypdfPp/U8QCXLRCdtERM+A318K9xT1Grpl07RMCjcxREYNG/HHrkJZCIWsz7hiMteJiaFqDLHwkdCIXSvFjP7TEajPhiJkfM5NKAOsuDl9/j/ShduX0P8XuJ5XpktHZSzIa1KIh7YVC1mYilzIbny+GuBhEbOCag5KSEeoSmiJWP76Y+fcTrjvijxgnAO0z44hczA5tm7qZoJC1kaVayvjhyeDeSHB/2z7Tt3MpAO27lpOnvk0RawCImVub6Vu2MlLTyJgMSRcbL0OXQVEt5qwh7YNC1i7wx29dyrI1Fl9LCZcSMRtN3gzVUiJtADVzpDFwz9DnWlxnjdIJpa3BlI4pCwvBfeDZofayPJo7A//tgkLWJtQay/TEZ+9LW0DrtiBfzO3jXmvc8F3aTf7oLw1ZHed//JdRp5FVsTJjJA1jcOyu8pgAF56L3cdSe8x5Q9oDhawdVKwxfZuHEJfyXNmlRHfPrtvjfr514mubMtl1vcC9e/P4V5bVACtdW94pPYrAxYyrxUTgn1ZZe6GQtQGNp9QM8J8tZ5j3VVzKEIjzoKE0WRtz5582kzZepvhjFqANJmonMQJVHLTK2guFrNXAGrOFIlWj4TFcHcS9EHDOBBqOo4DBkkAvFmR9wL3MXTkbtHzTXYPi3uO5xI1dQKusvVDIWgxETGoqu2unWyDAj26q41yet61LSdYPuI1wMX00XoYa43SmJ7KUQ9Aqax8UshajeWNxsbHyW/+s6bVxsUxg2DaAUY3oUq4/cDGn7X0NjcyERNle+2KBJRyXjlFllZGWQiFrIaV8bsXRrBHgl0Frt+8LujmosbzEfLGmAUsXffT7eWWg96obJR1Du8kOgaH3YJUx27+1UMhaCEb5liTKTLjHBLg3sMgwhJk2Q/JHP7pkYzmspWweSGe5fOo70feqfszsc0PcEvcf8bQQcD9hcTNBtrVQyFqFBPkXoj/0ZYttgZnRmsrdB5YtAzJq+Ms/MKS5TJ76VmSVAVfMIGQgd/lMeGN5UfXIaPAM+rcOClmLkCC/qbgeoeXWGkP3zHApdQBZvyDNnXuG1lgLwLNAky/gZ/3DKusZGZNnFZftn6pY3HHLyfpDIWsRpfzCikF+gORXWT8QH7vEdIuWcdkKWaEyQIs/IhNqMJFXtlDpRtsHQX88a4mJkpZAIWsF1sWAqxEXG5P2km8+J0F+DJYRWg6XktZY64BVhn7/tYIFqJhlB3eJ1ZWbPBO7Pd3L1kIhawH1uJX5uYume2SPfK9q71f5ZGys9Uy9snTP/ZQMDfoXYobVW3IvaZW1AgpZK0BVPEY4CriVKBS5yXGZR3zM/V2tgfz0eclxIq0Fo04hFgb80anQ/hIsVJ6dTyqdLbuXxbwhzYeDj7SAonUxUqnwrUYBQQ2YuJV924JdyrjtAJNCz/A7gr8vXHnDJImZM0/aF8z+ZR0wdtkQAJ7ZItxLxMz8DhplaL5MOeXGDBrSXChkzaZYkDhJqqsvuBhBYwT60ctFKAkTzPzuSZMkuq2I3fS5HwWX/frRe0wuQWIG93L0wH+JAv6uy49QAGKbRVuRExp9HKGEUm7RPn77/Dmgb1Ph3W0ysMaAuBoBNLGyqxIfc9/qKDgLEy9LkiZpD/lKk6TQQC9dlYqZuEoYWGSyfpFpGM2GQtZsKjGSuPiYipQWCr8L67mzjI21G7iX7gtG5zVOVqtHjHKcrGBIc6GQNRkEe+NyxwBqKzP926JaLuAWmtmEuZUbETwDf8QlgGeWsXHNWhZzSoSM7S6bDYWsyZSFLBO7HG9zBPn9pEvZtuJakvayMFGuuQyJWc/w1TLaEuJkIeQlRous6VDImon+AccIGawxkOkbDcZgclbEihxqrO1o87EQqV777GSdmOeU1jgZE2ObCYWsiWhsxA/0R4OILExLIejq375sGa2xziJutKUuhAXsZ352Irg8VSlipRLdy2ZCIWsiJaNv4WqLTN3IwvwEVgpaZJJfdpFC1inkYl4q2sBfrWufVKbyEmNTpaZCIWsmkUUWvs2IraSy3VWBfoUWWWeBFItQHFOeXdpOcT1daJdNjJM1FQpZM8FbOBV/i5H1DddkWVZ4hbjhx0jryU28EqyMARibND83EbstXmRL1jlpBhSyJlJCRn8qE/i9XAAKVqgwq296P6s/R4usY/BTLFxRy/QM2lrLWo3DbTErlgxpHhSyplLbnSguTJlsYKRxCFtcrwqkfWgG/3L3smeF2mW8qOhaNhMKWbNJp2ouTlWCxb5rmZ85b0jnEQoDIOAfsqyVWnmEZH2gkDWRcu5Q+BbjDY4/dy0SfrJlkRZZx5GfWh7wF1ETEUO8/8rSbz6stWwqFLI2IW9wI2UgmH+0JHGkkwgG/Pmo2g6FrI2gDLiD8HpLDek83G581DpDv2SV1pfyf/jFRJoJhaytVP78U3ylJ4LU8hBA9Owcw4zPs/VQyNqE/LGjdjI3tTzmQjqTymPyxaxkn2E5VEZLrF1QyNqE/NF7f/ehGi/Sefgvm/JzMybUuy9pDRSyJlKre2ONjbm1lmln/czALkM6C21XuQyJlyExdrjy1bOqpUKARa2Z8O42m0AbO/cP3U185Zu8s8F4lqFnhFQZ/TkUGigx9aLpcPCRppKp2cYOtV3FQGNj6Q9+aLdJKsinGv/hX8UsS25+HJ6X1lYqmMfYlWhvGfUcGxCzUFM1sn5QyJpIaoX+2tOZLumTrLxuqqogYB6jjidx4BGI1cVTf2c2Et2jN0S1lG5tpTyzYk7cTl/klrB/A2kWtWZC17KZIC5Sw61AX/0FqfEK11R2j+41pDNQoXLR74szb0ddMYVcTxkOjkWtqfDuNpMa3RyjEKTtH3+psBjs7x3Ls4PJdS83Gt1b9wZrK01xUcYmzfaNhjfUFxnbWzYVClkTievmWNMs0n3bjPQjW+k5wXdZNoJFlrU1edlKbV6ScV1LBfOLs2/LfHZge3C7qNcL1lo2FTruTSSum2MJEFsrDZ0qgtzUOdNXGUnJBVZAUukZfofZ85G/MUPX/L58n3zlcfP6E//DLCRolHEXCBnwxawwW+7iOi41Ixq3gcH+psLXRDOp0c0xcsYylb7I3J5g3RgLCk9s7lKHc81dRyMRAyN77xZhSyJ4BnHWsfYMm+kdDcc6YwagIesLhazJ4A+4WAy7lggQZ/u3V9VM+jVfagkkCQj0yN5/s+x3CFu3tdSShj4Dv8E4KFghwzWF+vMHGNeU8bHmQyFrMvImrlFziSRLDEKCgLGfggH6r/1DkzQyNWJiScwj67/2zmVpF4J9QeVtjCzjBPqXVQjYddI1Rpon6wOFrNmIkJWkdlJx/9i7KjWT6EbZzVNS+mwhSho5Gwebev3/L/sdv2nng0mid9f75NMXqdyVs/LZFVO7XFJLnG5l06GQNZl0lF+Uj35zhUoz+BcrvY8C9xPLYbUljfF//KsqMcM8fksaSIGBaxlKds1NnpF2lpmY51NifKxl8A43GwT8keGPUXayfeWfHKsrne2RN/riZVsorjm4bDkYfNdHzeSzXzdJAlbZS9/5jLiZGGAlqbWV/RWLONT8aHH6rOm2z07yAQN9+ZeQH2ifPYWs+dAiawHlP/R8VaysKqA/skdiR4ten/BRnGzPYZNU4EomVcTA0O/98TJLGSCuWVycMV322emy5QmzhaUUHNJUKGQtIKVxMicNw+27qmfrdTLvxsnc4DLyyXp2vs+Q1oLYmLr+vpjpKPDdW/ZUPS9F4mMY1zTdbUjzoZC1AIx7CEqBni5kedeAxGLm3nwuWMUPBt/1EUNay0DMPcdzWZgcl2ZJmufnu5alSrOzNIWsJVDIWgFEyVbBh7rs0T/+3q1j1g9bFPcS+AWj/51/GI2BSZoPAviITQJ9qUSNxO0zKuWmTd+Om2Mb/KNrHzxzk2YRawW8yy0iZYP6bhqG76r0bNsnCbLzb52q3q5SgPDmH7bxGtIaRm75bGy3POpWZod2xSTBVtzKbI8hrYFC1iLSMe5lFNi3Itaz7QazcGlcCkKoyxgIGa2y5gM3v2/XAZn3nwOak81ffMn0bN9n0l2D4R5h6Va2HApZq3Ddy0rtpWuVSdB/ZEzm551YmQua/gy/+z8a0lwQG0PuXqi2cu78MzLOQv/uAzEjipfKbmW2m25lC+GdbiHprv7KH3r5je3HwVB4uqy7AiFDH2VubEbFbvg9/yk2AZOsHVhjW977mWW/S48l1hrLTZ2VdWCNuc9PP/VFpRU8pDVQyFpIKtO1lBxrlndvjQkuDd7otWJlo7f/uSHNASLm9zmmLFx8WYL8A1ffFi3zY2TFxVl5xmkKWUuhkLWYdLZXYmBu20sQ5YwNX226h642cxeeixqS+ykZqMFkXtn6g3atg3s/uqyWEiBhGUH+LvtsXLfTRYP8YnmTlkIhazFpNFOyb2x5cztoZ4sA8RcMaAEXM47tf3CUgf91ZvT99y77TQVt4a3npSff/t3vC64DirlZiYUyyN96KGStBn/ojlXmW1wAb3zEYWbPPRN1uuhbZVg+8t7PGrI+bLnls1HjfN8iwzOYe+s5SZHRvseWYS0x6bIHLiWD/C2Hd7wNxFllQAvJ4J7D0qB8+syxYO0ZGH7PH1sX81ZD1kbfNXdW5Y359xk1lbKetca0U0wlErvKs0xn6Va2AwpZO/CsMheNlWV7h03vjhvN4tR5s4DuYky5e2y/sG3/g//O0ZbWAGqAt/7+UuWJb/kiLoa8sX5bCYNBVGTQGM/iwnMs5uflmdIaaw+8620iZJVVtdWzBaZ/581W0IbMzJnj5WHHvF4x8JkZ2Gm23/UVQ1bHVUe+YrID1S6lgjFHZ889bTLdQ6b/6gPLRCzq7hq93qKmktZY26CQtQvHKsPb3Bcx+Z7usi7mEbtOzrqYx6vW0QKlvWOM3P5nhjTGVnvPZICXiqWrqKDNIUa5MG227Pt49LvvWuLZSU0lrbG2wjvfRqSaXqyymXKwWDtbtAVC5xGA7ttxk/RGqr1j+O6PNl8KJXKSMMP2Xvl9jblihhpj5I2hljJ2lHH7zKK8MaZctBUKWZuB2yLZ/otzUWFRF0a/D1xz0HQNbBcLoVAZfgz4hRA1bxSzlcF90uA+8FtQoJZy5vWnTMa69QPvuH2ZyC0F+OeYN9YhUMjaDLL90WC8mJ+rSsfw42FD139Q2vhNvfrjqkbl/ifFrDbu/ZGWEo6LDiBiV17+gcQmh/d+LJxqAWCN2WcGl1LcStJWKGQdgFhl1j1BcDk0KrkE9XuGzNC77pGCNvXqj4IukVpxFLMwuCfaBCkUtMcnXhS4xxiGDzXH7shWkajZ+fz8JAP8HQSFrBOAUHUNLsVcop+rCxCaLw1ee9Dkp86bqfFjwWRanaeYVYOY2HCgHaX7feb1E6Y4PyEtK3Cv1SpGiwv3PsszUpeSAf6OgCMjdAjo9iVd7BN3BbWVqcrIPLLMcTV7beC/mM9JWsCctdL6dr1vWWxNC54UXBuonvzl/zKbGdRODlU6pQzFxQBaUcy9ecoM2OA+Ui1C7j1ALSVdys6Dr5MOYqkWc9qkHEsMwuRW+6OgoRNGBP+lf6xUdb9mrquJ2sydH390U3b9g2ve+bFHgyLmgns4f+EZ02vvaX+lZwtdr2pdsZhnKrWUA4Z0DhSyTkIy+kcwY/ILk+V+rWIsCDRhQtu/OVhmFTFzM//dGFDP6A1mxz0PJ3LU8tWCpls7PviwCD5w7w2IcsXsvYM1hns0NHYkWqYvjihnTONilmzPiDwr0jlQyDoNzUmyIlaoNBgPgYI5NHbYdI/uK7uZ58rtAV030xU/DAK8/chXxN3cyMCVhiu580Nfi4Zyc2sm3Uks2nNlS8wVMcXdrlipiBFLjHGxjoMxsg5EYi+VwH8xN2PS3QPBeA3mIWZly+Jp6b+s/5o7omWhgTFQAYCunCd/8T/N3Os/NRuJnh23mtEPHK3qxULx56fPPCntKCFiw9fdtexe+cF9dHaJFwzjYp0JhaxDKVtlhXLwH66i/e4O6usWNIhZpmfQzJ592uSmz5lh5JxZy8G1znRdINbZXX9jpk//P3Pl1980henzJslIw+/b/kxc51B+netSornX1KuPmby95gFbO6k9Wui6riULRMTsJMF9Jr52LBSyDibdPWQL3lJKhhYkLXBuUybUXpqSETfzyks/MMP7PhYNHutuA7SwYtxGTJf/5etm5tUfJk7Q4EZiMJZBO+Fa43LDdMrPXiznidmA/YC1XNH0y7XUYkXM1iCn2YllR0Mh63AyPcOmsHA5VszcT1gXWRsXmh4/Zi49923rZh60hfXGsnvk5EL5bHnvZ83A9R81C2+etKL2jY4XtCUB+w8i9vJbwI10Y1wYA0HjiCP7Pi7xs5AlpqiImVQ2OgbpXChknY5k9W9ZJmahzv1Al3WzYI3BKpt74ylTmLso1hosCjfPTNECjIKN2FL/dR+21tk/mtnTPxRh6yRQE4n4HrLuU5X0B1+09LfIlbSxrenfPSmN7rsGtpmh6++RVhLutu58FBOriFi2dwtrKBMAhSwJBMQshf7MKvgxMDR5Gr35T8zM2V9JasHi1DnJVkcqgm+ZuakGKgDqcuZtvG36hf9rZl//SdusNPQV1n/9hyX+hS539JzdT8V3J3OT4xLUR1ysb+fN4k7GVQDod4pYMqGQJQUVs9xUOW5jzLLgsytG0jFjRbxgnc1Yq2TRCpNrnYVcTfd3jBi09f1/Ln2d5dBT6oWTZv61n1qBeEWGRWvKZdpzQ/9q6H667513Rp0elm9BquanzkvvFVbAFmfOiagP7PmgjEzlp1b4oo4aYlSuYCBluPQUseRAIUsSImbD5QJnxaxUzNuCiiB0OLMfn+ieeetNn5JKAEzoY6vnqhtN/86bIvfM39atFNBlsIa6rMAMV7LkcxOvmNyll03+0mmZR2uExUuvmEaASHWN7rWfO03XyF6ZxzFqWVpKKLCPXnQxjB5iYalsj62VvE26C8fYB7peqOYXya4le/6SYmHXZUwseVDIEgjyykzFDUK2uWSaV2ow/eZKCqyz3m3l5FkEvhcvnymPCjT6LrFaXBFzrZXQPsvCtlcm4AoerKH8zHlTRPfPAbSpFFJAQlahnwLhr+PGv6L1Cwv2mp4382+eEjcS8T4kuCIWFsqlq74/RXHZtRE4UyySCYUsoSy1y7RitnBJMs5TzujWrpBFsTNbsKVpk7WupDXAeVhoL4mo9Hk9oaY9YfQJuWkyD9ewe69ZCT+epcIZEipf1JYC+Qtm7s3nRZgR1IeAac8VJnCefqoFtoElWT7vochyI8mDQpZgJEkz0y1WGSyhdLZQrgTwrDE/MRYFHYUeMTMk0SLDHRPiaRC5uMx4V9h8S8fNaQsREkXfPQyJpy+Y+L5oKx7mrQuZs1YlvOqyEFsBs9fkn4cvXpUZU7KxsHKycTqyaElyoZAlHRTE3q3l2jYUTGuloFJA2mxWCmdo9CUAAcDAGhBBcTmtdYYYmgTct7yzLGr926J9hFw9d5nu260F1eXpGKHwz8nvsVUpzFwwuStnzVzF+oL11GfjfN1b9ogo+wLrC9hSPCxfHvWoMmCItJ1kUD/xUMg2ArBmEDdLZ0TQCvOXysPNoV1gqlrMdL5KQKxwweWEe4lOGzHIyYKNOWGCqEHwukf2SI4aLEDdLiSQfmP1ldzTZe6iCqYVZIzpCesLlhfEFmtAtHAuvdv32R10BS3B0PnIOpEVhvs1bO9RtyEbAwrZBkJdTdc6QywNsTPfYvEtLHwi6J/ZNiQikZ+/YisEfieJpKiNRPoFtsz2jZpM37bK56h8pmrEluKEpspisxYWevqAAKMZEVzeojPICtJAenbdbF3id5q0k8waV1mgRNdq70N5pKpSpbnREK2wDQaFbKMBlxIpGelspfcMG8xOzZYD+eh5NibG5f+O/uqzvTdJ+kIpb60jWxOJxFoIDdIucpdEF8p6kO42XXBB7f7TmZ6oXaJ1MEU0JJjv1GLinIoQl9yM5MVByKLzwLHtvnqs2whLULriSXdVxc/cYfP8869KscCgxotz8on7krHXhHtANh4Usg2KdsVcts7mrZBcEctMLDTrgkbrVQQiFFvS76muXolFYQJYFxZT3opQYdZ+zl2UHKyitagWHcFSqUy585UvsKxgHfVaNxEWXbZ/uxWaEbEK0/Z4fjtI3x2uVRtpL7iSZ7dYdiOZVrHhoZBtcLQPLRW0grWExDKzMTQMRRdXwwf8pkvueumB7SZrJ7P1uqr1dXnBuqaiWm5szpjgYLfu9qFKAT/m5q8ftsAqAlapxSUbGwrZZqDibqJglzsJXDClnBUag9/7ZaCTUK2jfvcD8W6jc13H/QR+LMtft1YtqH9cXT+0XM4FbqqdYIlRwDYnFLLNhApaqb/sCkYxNCsKNs4llQLZpVpJxa+J1HyuUJqF4rqsuq7/m7uef8zQuq4o4vwRW0P8Tmw9CtimhkK2GZGBZcsxtFLeClphvlyzh1jaYqbievZEgfEq961UWiYw+ntI0HyLTn9zt3PX1VYCbqa/nAOsLxEuOxUXIpcV5yhWJYP4mxoK2SYHFlgGVhjGCKhYOBA1TGLZoGNB5I6lM5FY+PEy180LJeGGWgX4VpqibqssR6wLgXvUahYLRqsMRGi7uu0p9dL6IgKFjJSpWGlGBz5R101qAGeW1pMUDohaWlI8DObt5Me8XNHyrTe/KVLJHk/iW6WCdO0tAmaKSxUF2B9qXDPZcntSihfxoJCR5biiZqAneasxeREbCFupgNG2nQGEK9uIoFXiVSmDFI/qAD2+SfaFFSwMQCzrloqBw2dRWyBpIuVkXraDJLWhkJEVSVn3MpWt/lOBuBkZAR1WVKEsSLCy5BOzizE7S5d7TxNLLlX+ni5bdbDwKFpkNVDIyKqAuMHoopNHOgG+/gghiYdCRghJPBQyQkjioZARQhIPhYwQkngoZISQxEMhI4QkHgoZISTxUMgIIYmHQkYISTwUMkJI4qGQEUISD4WMEJJ4KGSEkMRDISOEJB4KGSEk8VDICCGJh0JGCEk8FDJCSOKhkBFCEg+FjBCSeChkhJDEQyEjhCQeChkhJPFQyAghiYdCRghJPBQyQkjioZARQhIPhYwQkngoZISQxEMhI4QkHgoZISTxUMgIIYmHQkYISTwUMkJI4qGQEUISD4WMEJJ4KGSEkMSTNYS0kd/+9rfm/PnzZtu2bWbfvn2GkNVAIVslFy9elAKIz7m5ObO4uGi6urrM8PCwTNddd53p7++vuY+f//znNZdjf7t27TKjo6Mr7qsV59sMLl++LOfU19dnWsHrr79uXnvtNbnmG2+8se71ca9uv/12QzoTClmDQABeeuklsSRCy1AoMWH5NddcY2666SaTzYZvM9ZbCYgPgLWyGoulkfOFmNVTuJPM7OxsdM0Qp5Xuqa7fDJGHSGL/sEYxkdVDIWsA/NGdOHFCPgH++GAx4RNiBUsHE4QD6+APdWJiwhw8eLBmQYDgXXvttVW/qchcuHBB9oV9ouBBbJp1vhAzfN8slgeuG5YZ7kk7gKWHZwwxpZCtDQpZA7iiAMvFFxUVKwgTCokKxLPPPmsOHToUu19sF/pDRgHDHzm2h2V26tQpEbw4C8/nl7/8Zd3nCxHD9J73vMdsJnBvIWbtcKvJ+kEhqxN1A0BIFHwgQLCgID6u69Yo2Mctt9wSuZh4i9ezH5zvlStX6j5fLI9bB9Yhzh/7wwQhRcGH0EIEa4FtcS64B7othAOCXK8Vgm1x3Xo9uCfYfi2xQ+yjVCrJ+eEFdfjw4bpfED6ha6x1jhob1evRbRU8bwprY1DI6gTWFUDBrVeQsJ4G2PHHuhohAxqUxx8+guPNOt8QEJDnn39eCqsLzgXXhuPEuc5Yx7UK3d/ruR84JrYPxRI1bgULcjWuIe4pttXz+8UvflHTao4j7hprnaN/Pdg2tD2pHwpZHeAPT//Q/FjWSqCwYnsIEPaxFguiXvRYoNHzdYHYwPUCENLrr79ePiEwKMCw0jQO54sZfoflobWjuA9amLFtXAWEC7bHurq9Wn8a18N9hYjAmsJ5NYq67rovfDZSoRJ3jfjuxh5xjhBJtUDvvvtu+cTvuD7fGqY11jgUsjrAH6XSaFDWLWAI/K/mj1RrzsCWLVtWXN8933rWjzumWnUonH4FAO7Dzp07o7ihHwfE7yjQuF5f5HBPsK0KVQh1jUPbYx7HgtsOMcTnaqwpAOHC/YLlietFGki94l/rGoF7jbg/6r7qevpywifFa20ws78O1mL2r8cfqGu5oHCsRD6fj+ZXG/eB6Op1x1UA4NoQzwGu1erOx8V7VsrLUhGF0MTdQyzD9Wk6xWrB9ekx4EbX87wbvUasq3FOsv7QIquDtYiRG1uKcw+xjl94IEZ4k2sVPahVqOPAflYjZufOnZNPWGO1jqmpHDiOWpwax8P11rJgtbbWFyH/ftQSqe7u7uherTaFAecJi+r48eN1B/8buUaNb8ZZn2TtUMjqwM06R6FqpMC4f7xxBUNTH2qB+FS98ZuhoaFoHgVuNQVcrbp6RFDFRMVHt60nbhXK6HfF/+TJk6Ye/MqIRoHgvP/97xdXsJ7gfyPXqEKWy+UMaQ4UsjrAH6JaHdousF5gUYG4XDGAt7pvraFgauF0A8X1gLiYni+Ovxoha8SK0wLqX4Mbq4tjJQGC2wehXInVBPt9tL2nBv8Re1upkqWea6xnHbI2KGR1gD9mWERa04ZgcD0FB6KHoDWolW+FGivf2kIBf/LJJ6Oge6PBbD1fHL+RnC0FYohWBZhqAYvPt070WDh3WCK17lXIbYToqxCnUqkVc9XWEzf4j2cdd+7uNa7kvqsbun37dkOaA4P9deI2qtZq81qggGrqArZrNJfLDRRrQu1qzxfnUc/5wq1SUdJKBW2rGcfLL78sn67FqRYsqLUtlrkVEy4QYqApHnE0I+4EK1AFLG7/7jXWcn/da0RybIi4e0DqJ4P/7rzzzpFSqfR5zPf29po77rjDkGoymYwU1DfeeMMsLCyYM2fOSAHzq841Hwm1X8ViMaqax3310UIe12i4p6dH9v/WW2+ZyclJc/XVV9edT7aa84UlMj09LcfB+ULEcFy1mhDP0uOrUOPcAGruBgcHo2NjwjLsD2IwMjISbYv9vvDCC+b06dPRefhtHvEdFQ44Z1iFmhSsYJ+/+c1vZD87duyQe1UPboPxuJcLzv2qq64SqwzPEPjr+9eI88T5xV0jLD0/cVdbAuD54MWBbbEdvjeSN7iZeeqpp8z8/Dxdy0bAHypqszR3Cm6buo4hICRrbW7itg6AJfiBD3yg7vjVas7XTbXQnjBgFWnbUVyLG7/T9fxCivNWaw7nj0ljgWphYV+wUkLnpDWJeu6wejD5x9f9rUeMzMUN/sfhXqPeW/8aQVzPJXD5tekbwgh4UeBlgmtZbV7cZoUWWYPom1kLVCiQC0HYv3+/FPBab9aVLDJ3f7AOcKxCoSAWyHqdL5Zv3brV3HrrrVH7UBccy90Wn7BSsB6WHThwILaJEM4bhRIWBrbFdpoFv2fPHtl2ZmZGRDrUCwXWU0tFj+0eH/vAiwLnXy/1WGQKrlst4rj1cY2I4ak1q9eoy3Bf4xJsNTUDVi/uESZYebivsAjJyqhFlsKXo0ePjtkCIkEYuAD33nuvIfUD9wDWgsZTQm9gTaBcS5Oh9cINUDdqLbrW1GqOu9ptgWuJdWomvF5jqCa6FnptzPBvjEceeUReBHQt1wF1BbS7Ha2+VwtDXUPt8bXd8Y+1FJZ2bQsaFYd2sJa2tIyLrR4K2TqhtYya3Oo3m9HmPPxjJWT9oZCtM9qTAawwdTU1VkQRI6Q5UMiaBFzIdnWhTMhmgwmxhJDEQyEjhCQeChkhJPFQyAghiYdCRghJPBQyQkjioZARQhIP88gIWSe0pwx0FrDevXGQ2lDI2oCO3K39UWlDaPzxNzoKN+kctEnaWscPII1DIWsx2q9XCB1pB31Uoavp2267jb0h1IE2B8O9amW32KRzoJC1CLyl3QFptX8rtcJ0FCLt5x/9vOso2qQ2es+0bzCy+aCQtQiMyKMihv7oQ4PeaueC6MsMIoZO+QghK0MhawFuF9Nx3R67wEVayRJDPEb7OYM1h26SIYTuoCM+GozW7oQQp1N3FttjgsiuFKiG5agxvkaPjUA41tdjw1LVbp0xr9Yolq20b3056AsCn27X1LgWf2R2PYbGJ9G5pL5AVmrkv5p7TloDhawFuGNb1jvIbi1CcTaIi462FCeWGoxGAUYnkG5QWns2RSGvJbbrcWxYm35vsa+++qqMyOQHymvtG/ty+3zDtu53vzdeiJB/3bofdU3RT39oTIRa141tQxY2aR0UsibjFq71FjG1OHTQCggmChWW1+qTXoepw/loP2nYHtvpOJqhcQTW49g6dBr2DQtIRxqHhYN7hd8hQBiUBL9BZOLOCx1ZYhtYZhhpSQd7Udz+33TwFj02rh3Hdo+BdUIjjLvX7Z6fe916T0l7oJA1GXdcxLXmFmlhBnDRXLHQcSVR24mCjfVQ4ELWBQo/XFf/fCBMcYMC13NsDEALoap17ND2AN9D4olzxHk99thjIjoqWEC7h3YFK87FU6GBeOp4oe4xIEzHjx+PLCytNHCv27cI9brxG7bl+JTtg5n9LaTeYdy08PiujH5HYYyzeFQQIFbq0vqgkIZEFYIwNjYm8/7AtK5FEnds7FePDVcxRNy549hxuXNYpm6ijtrdCDrwC4hzASFKOiiwe9/0umuFBdxtSXugRdaBqJsH/JgQgEvkxoJ81F2LK/S7d++O3VYtGj+OpMPIQQDrOXbcCN0rJfrqsHP+9qFh9+pFzxfWKvYTty+9dvf69DxWqgiAiNcaVZ00FwpZk9GCDSYmJtZUu6Wuiw5wshJxBbaWZRg3roDuq95jx2W3x7nXGuuqNYDwalFrDMJea8Bdfxs8K73nK1nTrLVsLxSyJqODsOLNDpdlPRI2sQ/Eo1aiXle2Eeodi6DRY+tQeu6Awi46WvlqUHHGc6jXBWxUmNgsqb1QyFoA4juwNuCyQMxWO0jv0NBQNLhuqzPYEQzHsVFg1/vYsJRUpFDBELLatNXDaoBLCWBVQoQbEVm97lruNEAlBGkfDPa3AG2KBJ5//vlVu09qTWhSZhzqSq0nKr7+eJ3rcWy3ti/OEqpnv3G1hqj1hHhBhF988cXY7fUl4eJed1zlCfbL+Fh7oZC1CFT5o5Dijx4pCphQOPykVMSf4nKS3PQExHo0v0rRAvX444+ve8Gq99hI32j02K4FFqrtXCl2ptvjfNxzctu17t+/X+Zxf3Hv3coEzfU7ceKE5JG5uNeN54Jrc58ZtsP9aMbLg9QPXcsWARE7ePCgFAbNVapVOFH4QtX9EERtfK4pGmrFNLsw4dhIKsX5r+ex9Vp1n7gvcOkAXDYIh8YZQ8BqUoGBiEN4dN0Pf/jD8gmrWMVW773moOF3FSe13NxKDyTZQuTctBh9Kel2sJZhKVPQ2gOFrIXgjx8xIBQiuCkhF027okHBC9Ug4jcks2IfsF5QYLXwYBkKPEShGf2ZYf/++bsFV5NDV3NsbIdrV0vPbcKEZTi2tgoInRfuCUTWbbaE84CrqDEx7EfTJFQg3b7gtMG+D87h7rvvjlou+Ofnji5P2kMK/x09enSsUChInfrIyIi59957DWkNKBAIQiNNw89Sb2QfoB0pAM04tgrMau5HI9u6L4BGjrOW8yPryyOPPGImJydpkbUbCMBaRaCdOUzNOPZaBKKRbVd77hSwzoPBfkJI4qGQEUISD4WMEJJ4KGSEkMRDISOEJB4KGSEk8VDICCGJh0JGCEk8FDJCSOKhkBFCEg+FjBCSeChkhJDEQyEjhCQeChkhJPFQyAghiYdCRghJPBQyQkjioZARQhIPhYwQkngoZISQxEMhI4QkHgoZISTxUMgIIYmHQkYISTwUMkJI4qGQEUISD4WMEJJ4KGSEkMQjQjY/Pz+pP9h5QwghSWBysixdImQPP/wwvskvEDJdSAghncr58+ejede1PKkzL774oiGEkE7mwoUL0XwkZKVS6fs6TyEjhHQ6J09GtteSkOXz+W/o/Pj4uEyEENKJIPx15syZ6HskZJU42RP6/fjx44YQQjqRY8eOuV+/UZV+Yd3LL+s8LLKnnnrKEEJIJ3HixAnz7LPPRt8zmcyXM+4KP/vZz8YPHTq0NZVKHcT306dPm/3795vBwUFDCCHtBjWV3/nOd9yfvvbggw9+K+OveMcdd5ywCvcRO7sL30+dOmW2b98uEyGEtAsE97/73e8ini/frcE1/tBDD30U88uEzJpt84cPH/6hnf2knUawEcQM7Nq1y2SzWUMIIa0Cua2PPfaYefzxx6tELJ1O/9FPfvITSXpNxW189OjRsWKx+M82bjamv42MjJgjR46YPXv2yDwhhDQLCBji9JjcFkcqYg888MB49FutHUHMrALebzf8z/6ysbExs3PnTtPb20tRI4SsC9qyCLEwN73C4YmBgYF/d//991c1P0qZOrjvvvs+Y8XsS651RgghrQJWGLIqbEzsG8HlpgEgaPYD1tldhhBCms8TdvqmtcK+51thLv8KHdJLJwWJ6/wAAAAASUVORK5CYII="
 }
