+++ Tests/InjiVcRendererTests/InjiVcRendererTests.swift
@@ -52,19 +52,6 @@ final class InjiVcRendererTests: XCTestCase {
         super.tearDown()
     }
     
-    func testParseVcJson_Unsupported_CredentialFormat() {
-        let invalidJson = #"{"name": }"#
-        
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : CredentialFormat.fromValue("mso_mdoc"), vcJsonString: invalidJson)) { error in
-            guard let vcError = error as? VcRendererException else {
-                XCTFail("Expected VcRendererException but got \(error)")
-                return
-            }
-            XCTAssertEqual(vcError.errorCode, VcRendererErrorCodes.unsupportedCredentialFormat)
-            XCTAssertTrue(vcError.message.contains("Only LDP_VC credential format is supported"))
-        }
-    }
-    
     func testParseVcJson_Supported_CredentialFormat() throws {
         let vcJsonString = """
         {
@@ -82,7 +69,7 @@ final class InjiVcRendererTests: XCTestCase {
             }
         }
         """
-        let resultAny = try renderer.renderVC(credentialFormat : CredentialFormat.fromValue("ldp_vc"), vcJsonString: vcJsonString)
+        let resultAny = try renderer.generateCredentialDisplayContent(credentialFormat : CredentialFormat.fromValue("ldp_vc")!, vcJsonString: vcJsonString)
         let result = resultAny.compactMap { $0 as? String }
         XCTAssertEqual(result, ["<svg>Email: test@gmail.com, Mobile: 1234567890</svg>"])
     }
@@ -90,7 +77,7 @@ final class InjiVcRendererTests: XCTestCase {
      func testParseVcJson_InvalidJson_Throws() {
          let invalidJson = #"{"name": }"# 
          
-         XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: invalidJson)) { error in
+         XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: invalidJson)) { error in
              guard let vcError = error as? VcRendererException else {
                  XCTFail("Expected VcRendererException but got \(error)")
                  return
@@ -102,7 +89,7 @@ final class InjiVcRendererTests: XCTestCase {
 
 
     func testHandlesMissingRenderMethod() {
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: "{}")) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: "{}")) { error in
             assertVcRendererException(error,
                        expectedMessage: "RenderMethod object is invalid",
                                       expectedCode: VcRendererErrorCodes.invalidRenderMethod
@@ -112,7 +99,7 @@ final class InjiVcRendererTests: XCTestCase {
 
     func testHandlesInvalidJsonInput() {
         let vcJsonString = #"{ "renderMethod": [ "invalid" ] }"#
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
             assertVcRendererException(error,
                        expectedMessage: "RenderMethod object is invalid",
                                       expectedCode: VcRendererErrorCodes.invalidRenderMethod
@@ -122,7 +109,7 @@ final class InjiVcRendererTests: XCTestCase {
 
     func testHandlesWithoutRenderMethodField() {
         let vcJsonString = #"{"someField": "someValue"}"#
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
             assertVcRendererException(error,
                        expectedMessage: "RenderMethod object is invalid",
                                       expectedCode: VcRendererErrorCodes.invalidRenderMethod
@@ -132,7 +119,7 @@ final class InjiVcRendererTests: XCTestCase {
 
     func testHandlesWithRenderMethodAsEmptyObject() {
         let vcJsonString = #"{ "renderMethod": { } }"#
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
             assertVcRendererException(error,
                        expectedMessage: "RenderMethod object is invalid",
                                       expectedCode: VcRendererErrorCodes.invalidRenderMethod
@@ -142,7 +129,7 @@ final class InjiVcRendererTests: XCTestCase {
 
     func testHandlesWithRenderMethodAsEmptyArray() {
         let vcJsonString = #"{ "renderMethod": [] }"#
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
             assertVcRendererException(error,
                        expectedMessage: "RenderMethod object is invalid",
                                       expectedCode: VcRendererErrorCodes.invalidRenderMethod
@@ -152,9 +139,9 @@ final class InjiVcRendererTests: XCTestCase {
 
     func testHandlesInvalidRenderSuite() {
         let vcJsonString = #"{ "renderMethod": [ { "type": "TemplateRenderMethod", "renderSuite": "invalid-suite" } ] }"#
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
             assertVcRendererException(error,
-                       expectedMessage: "Render suite must be '\(Constants.SVG_MUSTACHE)'",
+                       expectedMessage: "Render suite must be '\(Constants.svgMustache)'",
                                       expectedCode: VcRendererErrorCodes.invalidRenderSuite
                    )
         }
@@ -162,9 +149,9 @@ final class InjiVcRendererTests: XCTestCase {
 
     func testHandlesInvalidType() {
         let vcJsonString = #"{ "renderMethod": [ { "type": "invalid", "renderSuite": "svg-mustache" } ] }"#
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
             assertVcRendererException(error,
-                       expectedMessage: "Render method type must be '\(Constants.TEMPLATE_RENDER_METHOD)'",
+                       expectedMessage: "Render method type must be '\(Constants.templateRenderMethod)'",
                                       expectedCode: VcRendererErrorCodes.invalidRenderMethodType
                    )
         }
@@ -172,9 +159,9 @@ final class InjiVcRendererTests: XCTestCase {
 
     func testHandlesRenderMethodAsJsonWithInvalidSuite() {
         let vcJsonString = #"{ "renderMethod": { "type": "TemplateRenderMethod", "renderSuite": "invalid-suite" } }"#
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
             assertVcRendererException(error,
-                       expectedMessage: "Render suite must be '\(Constants.SVG_MUSTACHE)'",
+                       expectedMessage: "Render suite must be '\(Constants.svgMustache)'",
                                       expectedCode: VcRendererErrorCodes.invalidRenderSuite
                    )
         }
@@ -182,9 +169,9 @@ final class InjiVcRendererTests: XCTestCase {
 
     func testHandlesRenderMethodAsJsonWithInvalidType() {
         let vcJsonString = #"{ "renderMethod": { "type": "invalid", "renderSuite": "svg-mustache" } }"#
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
             assertVcRendererException(error,
-                       expectedMessage: "Render method type must be '\(Constants.TEMPLATE_RENDER_METHOD)'",
+                       expectedMessage: "Render method type must be '\(Constants.templateRenderMethod)'",
                                       expectedCode: VcRendererErrorCodes.invalidRenderMethodType
                    )
         }
@@ -212,7 +199,7 @@ final class InjiVcRendererTests: XCTestCase {
             }
         }
         """
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJson)) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJson)) { error in
             assertVcRendererException(error,
                        expectedMessage: "Template ID is missing in renderMethod",
                                       expectedCode: VcRendererErrorCodes.missingTemplateId
@@ -242,7 +229,7 @@ final class InjiVcRendererTests: XCTestCase {
             }
         }
         """
-        let resultAny = try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJson)
+        let resultAny = try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJson)
         let result = resultAny.compactMap { $0 as? String }
         XCTAssertEqual(result, [
             "<svg>Address : TEST_ADDRESS_LINE_1eng****TEST_REGIONeng****TEST_CITYeng***</svg>"
@@ -266,7 +253,7 @@ final class InjiVcRendererTests: XCTestCase {
             }
         }
         """
-        let resultAny = try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)
+        let resultAny = try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)
         let result = resultAny.compactMap { $0 as? String }
         XCTAssertEqual(result, ["<svg>Email: test@gmail.com, Mobile: 1234567890</svg>"])
     }
@@ -299,7 +286,7 @@ final class InjiVcRendererTests: XCTestCase {
             ]
         }
         """
-        let resultAny = try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)
+        let resultAny = try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)
         let result = resultAny.compactMap { $0 as? String }
         XCTAssertEqual(result, [
             "<svg>Email: test@gmail.com, Mobile: John Doe</svg>",
@@ -328,7 +315,7 @@ final class InjiVcRendererTests: XCTestCase {
             }
         }
         """
-        let resultAny = try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)
+        let resultAny = try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)
         let result = resultAny.compactMap { $0 as? String }
         XCTAssertEqual(result, ["<svg>Email: test@test.com, Mobile: -</svg>"])
     }
@@ -353,7 +340,7 @@ final class InjiVcRendererTests: XCTestCase {
         """
         
     
-        let resultAny = try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)
+        let resultAny = try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)
         let result = resultAny.compactMap { $0 as? String }
         XCTAssertEqual(result, ["<svg>Email: test@test.com, Mobile: 1234567890</svg>"])
     }
@@ -376,7 +363,7 @@ final class InjiVcRendererTests: XCTestCase {
                }
            }
         """
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
             assertVcRendererException(error,
                        expectedMessage: "Multibase validation failed: Mismatch between fetched SVG and provided digestMultibase",
                                       expectedCode: VcRendererErrorCodes.multibaseValidationFailed
@@ -402,7 +389,7 @@ final class InjiVcRendererTests: XCTestCase {
                }
            }
         """
-        XCTAssertThrowsError(try renderer.renderVC(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
+        XCTAssertThrowsError(try renderer.generateCredentialDisplayContent(credentialFormat : .ldp_vc, vcJsonString: vcJsonString)) { error in
             assertVcRendererException(error,
                        expectedMessage: "Multibase validation failed: digestMultibase must start with 'u'",
                                       expectedCode: VcRendererErrorCodes.multibaseValidationFailed
@@ -427,8 +414,8 @@ final class InjiVcRendererTests: XCTestCase {
           }
         }
         """
-        let resultAny = try renderer.renderVC(
-            credentialFormat: CredentialFormat.fromValue("ldp_vc"),
+        let resultAny = try renderer.generateCredentialDisplayContent(
+            credentialFormat: CredentialFormat.fromValue("ldp_vc")!,
             vcJsonString: vcJsonString
         )
         let result = resultAny.compactMap { $0 as? String }
